Refactor: separate club apis & rename WishListData
